requirements.txt
fastapi[standard]>=0.113.0
pydantic>=2.7.0
uvicorn
ultralytics
numpy
pillow
asyncio
opencv-python-headless